Move collection endpoint tests into separate files
